+++ src/parser/roc_video_parser.cpp
@@ -73,6 +73,14 @@ rocDecStatus RocVideoParser::Initialize(RocdecParserParams *pParams) {
     return ROCDEC_SUCCESS;
 }
 
+rocDecStatus RocVideoParser::MarkFrameForReuse(int pic_idx) {
+    if (pic_idx < 0) {
+        return ROCDEC_INVALID_PARAMETER;
+    }
+    //todo::
+    return ROCDEC_NOT_IMPLEMENTED;
+}
+
 void RocVideoParser::InitDecBufPool() {
     for (int i = 0; i < dec_buf_pool_size_; i++) {
         decode_buffer_pool_[i].use_status = kNotUsed;
@@ -106,7 +114,6 @@ ParserResult RocVideoParser::OutputDecodedPictures(bool no_delay) {
             pfn_display_picture_cb_(parser_params_.user_data, &disp_info);
             decode_buffer_pool_[output_pic_list_[i]].use_status &= ~kFrameUsedForDisplay;
         }
-
         num_output_pics_ = disp_delay;
         // Shift the remaining frames to the top
         if (num_output_pics_) {