+++ CHANGELOG.md
@@ -48,6 +48,7 @@ Documentation for rocDecode is available at
 ### Additions
 
 * Clang - Default CXX compiler
+* Parser - Add new API rocDecParserMarkFrameForReuse()
 
 ### Optimizations